+++ go-client-sdk/sdk.go
@@ -93,10 +93,8 @@ type SDK struct {
 	// Generation - Endpoints for purely testing valid generation behavior.
 	Generation *generation
 	// Globals - Endpoints for testing global parameters.
-	Globals      *globals
-	Nested       *nested
-	NestedFirst  *nestedFirst
-	NestedSecond *nestedSecond
+	Globals *globals
+	Nested  *nested
 	// Pagination - Endpoints for testing the pagination extension
 	Pagination *pagination
 	// Parameters - Endpoints for testing parameters.
@@ -275,8 +273,8 @@ func New(opts ...SDKOption) *SDK {
 		sdkConfiguration: sdkConfiguration{
 			Language:          "go",
 			OpenAPIDocVersion: "0.1.0",
-			SDKVersion:        "1.30.1",
-			GenVersion:        "2.91.4",
+			SDKVersion:        "1.30.2",
+			GenVersion:        "2.93.0",
 			Globals: map[string]map[string]map[string]interface{}{
 				"parameters": {},
 			},
@@ -329,10 +327,6 @@ func New(opts ...SDKOption) *SDK {
 	sdk.Globals = newGlobals(sdk.sdkConfiguration)
 
 	sdk.Nested = newNested(sdk.sdkConfiguration)
-
-	sdk.NestedFirst = newNestedFirst(sdk.sdkConfiguration)
-
-	sdk.NestedSecond = newNestedSecond(sdk.sdkConfiguration)
 
 	sdk.Pagination = newPagination(sdk.sdkConfiguration)