+++ RELEASES.md
@@ -486,3 +486,18 @@ Based on:
 - [typescript v1.31.1] typescript-client-sdk
 ### Releases
 - [Go v1.30.1] https://github.com/speakeasy-api/openapi-generation-tests/releases/tag/go-client-sdk/v1.30.1 - go-client-sdk
+
+## 2023-09-02 00:11:26
+### Changes
+Based on:
+- OpenAPI Doc 0.1.0 
+- Speakeasy CLI 1.77.2 (2.93.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v1.30.2] go-client-sdk
+- [java v1.31.2] java-client-sdk
+- [php v1.31.2] php-client-sdk
+- [python v1.31.2] python-client-sdk
+- [ruby v1.31.2] ruby-client-sdk
+- [typescript v1.31.2] typescript-client-sdk
+### Releases
+- [Go v1.30.2] https://github.com/speakeasy-api/openapi-generation-tests/releases/tag/go-client-sdk/v1.30.2 - go-client-sdk